Split out data_parallel tests from test_nn.py into a separate test_data_parallel.py file
Splitting data parallel tests out of test_nn.py since its easier to
manage and track these tests separately and failures can be routed to
appropriate POCs.

Differential Revision: [D18011663](https://our.internmc.facebook.com/intern/diff/D18011663/)

[ghstack-poisoned]
